+++ Changelog.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - `-no-rust-enums` generate integer constants instead of enums
 - Derive Debug when possible
+- Support for inline enum declarations in struct fields (e.g.: `struct { enum { Option_A, Option_B } field; };`)
 
 ### Changed
 - Use `clang_sys` instead of the internal ffi